+++ src/views/Run/internal/ResultDisplay/StatCard/internal/StatCard.helpers.tsx
