+++ src/views/Run/internal/ResultDisplay/StatCard/internal/StatCard.types.ts
